backend/pyproject.toml
[project]
name = "beautyops-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
]
